requirements.txt
google-generativeai>=0.2.0
github3.py>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.4.2
python-multipart>=0.0.6
fastapi>=0.104.1
uvicorn>=0.24.0
aiohttp>=3.9.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
